src/main/java/util/validator/SingleNumberValidator.java
package util.validator;

public class SingleNumberValidator {

    private static final int LOWER_BOUND = 1;
    private static final int UPPER_BOUND = 45;

    public static void validate(String input) {
        validateNumberFormat(input);
        int number = Integer.parseInt(input);
        validateInRange(number);
    }

    private static void validateNumberFormat(String input) {
        try {
            Integer.parseInt(input);
        } catch (NumberFormatException e) {
            throw new NumberFormatException("숫자가 아닌 문자열이 입력되었습니다.");
        }
    }

    private static void validateInRange(int inputNumber) {
        if (LOWER_BOUND > inputNumber || inputNumber > UPPER_BOUND) {
            throw new IllegalArgumentException("1 이상 45 이하의 범위를 벗어났습니다.");
        }
    }
}
